ci: Lint fixes. Move balance commands and ERC20 abi to shared.
