+++ apps/extension/src/application/idriss-send/components/send-widget.tsx
@@ -156,14 +156,14 @@ export const SendWidget = ({ widgetData }: Properties) => {
                     Log in
                   </Button>
                 )}
-                {!sender.haveEnoughBalance ? (
+                {sender.haveEnoughBalance ? sender.isError && (
+                  <ErrorMessage className="mt-4">
+                    Something went wrong.
+                  </ErrorMessage>
+                ) : (
                   <ErrorMessage className="mt-4">
                    Not enough {selectedToken?.symbol ?? 'tokens'} in your wallet. Add funds
                    to continue.
-                  </ErrorMessage>
-                ) : sender.isError && (
-                  <ErrorMessage className="mt-4">
-                    Something went wrong.
                   </ErrorMessage>
                 )}
