+++ apps/extension/src/application/idriss-send/hooks/use-erc20-transaction.ts
@@ -4,12 +4,11 @@ import { getChainById } from '@idriss-xyz/utils';
 import { EMPTY_HEX } from '@idriss-xyz/constants';
 import { Wallet } from '@idriss-xyz/wallet-connect';
 
-import { createWalletClient, TransactionRevertedError } from 'shared/web3';
+import { createWalletClient, ERC20_ABI, TransactionRevertedError } from 'shared/web3';
 import { useObservabilityScope } from 'shared/observability';
 
 import {
   CHAIN_TO_IDRISS_TIPPING_ADDRESS,
-  ERC20_ABI,
   TIPPING_ABI,
 } from '../constants';
