+++ apps/extension/src/application/trading-copilot/commands/index.ts
@@ -7,8 +7,6 @@ import { GetEnsAddressCommand } from './get-ens-address';
 import { GetFarcasterAddressCommand } from './get-farcaster-address';
 import { GetFarcasterUserCommand } from './get-farcaster-user';
 import { GetQuoteCommand } from './get-quote';
-import { GetEnsBalanceCommand } from './get-ens-balance';
-import { GetTokenBalanceCommand } from './get-token-balance';
 import { GetSiweMessageCommand } from './get-siwe-message';
 import { VerifySiweSignatureCommand } from './verify-siwe-signature';
 import { GetTokensListCommand } from './get-tokens-list';
@@ -30,8 +28,6 @@ export const COMMAND_MAP = {
   [GetEnsNameCommand.name]: GetEnsNameCommand,
   [GetEnsAddressCommand.name]: GetEnsAddressCommand,
   [GetQuoteCommand.name]: GetQuoteCommand,
-  [GetEnsBalanceCommand.name]: GetEnsBalanceCommand,
-  [GetTokenBalanceCommand.name]: GetTokenBalanceCommand,
   [GetSiweMessageCommand.name]: GetSiweMessageCommand,
   [VerifySiweSignatureCommand.name]: VerifySiweSignatureCommand,
   [GetTokensListCommand.name]: GetTokensListCommand,
@@ -52,8 +48,6 @@ export { GetFarcasterAddressCommand } from './get-farcaster-address';
 export { GetFarcasterUserCommand } from './get-farcaster-user';
 export { GetSiweMessageCommand } from './get-siwe-message';
 export { VerifySiweSignatureCommand } from './verify-siwe-signature';
-export { GetEnsBalanceCommand } from './get-ens-balance';
-export { GetTokenBalanceCommand } from './get-token-balance';
 export { GetQuoteCommand } from './get-quote';
 export { GetTokensListCommand } from './get-tokens-list';
 export { GetTokensImageCommand } from './get-tokens-image';
