+++ apps/extension/src/shared/web3/commands/get-ens-balance.ts
