+++ apps/main-landing/src/app/creators/obs/utils.ts
@@ -1,6 +1,5 @@
 import { type Hex } from 'viem';
 import { CHAIN } from '@idriss-xyz/constants';
-
 import { clientEthereum } from '@idriss-xyz/blockchain-clients';
 const SELL_TOKEN_BY_NETWORK: Record<number, string> = {
   [CHAIN.BASE.id]: '0x833589fcd6edb6e08f4c7c32d4f71b54bda02913',
